MiniMetroBot/pyproject.toml
[tool.poetry]
name = "MiniMetroBot"
version = "0.1.4"
description = "Bot which detects main objects in the game Minimetro."
authors = ["TomGFFM <thomas.graf@icloud.com>"]
license = "BSD"
readme = "docs/README.md"
include = ["minimetrobot"]
exclude = ["minimetrobot/yolo_tuning"]


[tool.poetry.dependencies]
python = "^3.10.9"
comet-ml = "3.33.6"
ipython = "8.14.0"
numpy = "1.24.2"
opencv-contrib-python = "4.7.0.72"
opencv-python = "4.7.0.72"
pillow = "10.0.0"
pyyaml = "6.0.1"
screeninfo = "0.8.1"
torch = "2.0.0"
torchvision = "0.15.1"
ultralytics = "8.0.162"
pyautogui = "^0.9.54"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
metrobot = "minimetrobot.__main__:main"
